Use first 3 version bits to signal block type

+++ src/rpc/mining.cpp
@@ -680,7 +680,7 @@ static UniValue getblocktemplate(const JSONRPCRequest& request)
             }
         }
     }
-    result.pushKV("version", pblock->nVersion);
+    result.pushKV("version", (uint64_t)pblock->nVersion);
     result.pushKV("rules", aRules);
     result.pushKV("vbavailable", vbavailable);
     result.pushKV("vbrequired", int(0));
